fix buy order

+++ internal/app/models/vo/business.go
@@ -6,7 +6,7 @@ type BuyRequest struct {
 	//Base
 	Amount      uint64 `json:"amount" validate:"required"`
 	Account     string `json:"account" validate:"required"`
-	OrderType   uint8  `json:"order_type"`
+	OrderType   uint32 `json:"order_type"`
 	OperationId string `json:"operation_id"`
 }
